requirements.txt
# python < 3.13
geopy==2.3.0
pytest==7.4.0
logicsponge-core
logicsponge-monitoring